index: reset proof-of-work verifier less often
The `validity` is a half of the interval width. If the nonce is 1 minute
in the future, it will be accepted for 2 minutes (until it is 1 minute
in the past). Reset not often than once 2 minutes to make sure that each
bloom filter will have nonces that can still be accepted.